+++ weblate/vcs/tests/test_github_models.py
@@ -11,7 +11,7 @@ from unittest.mock import patch
 
 from asgiref.sync import async_to_sync
 from django.core.cache import cache
-from django.test import TestCase
+from django.test import SimpleTestCase, TestCase
 
 from weblate.trans.models import Component, Project
 from weblate.utils.tests import http_mock
@@ -23,6 +23,8 @@ from weblate.vcs.github import (
     GitHubInstallation,
     InstallationRemoval,
     exchange_github_app_manifest_code,
+    get_github_repository_clone_url,
+    get_github_repository_identity,
     get_installation_token,
     normalize_github_callback_code,
     normalize_github_installation_id,
@@ -32,6 +34,49 @@ from weblate.vcs.tests.utils import generate_private_key
 from weblate.workspaces.models import Workspace
 
 SETTINGS_PRIVATE_KEY = generate_private_key()
+
+
+class TestGitHubRepositoryURLs(SimpleTestCase):
+    def test_repository_identity(self) -> None:
+        for repository in (
+            "https://github.com/WeblateOrg/weblate.git",
+            "https://user:token@github.com/WeblateOrg/weblate.git",
+            "git@github.com:WeblateOrg/weblate.git",
+            "ssh://git@github.com/WeblateOrg/weblate.git",
+            "git://github.com/WeblateOrg/weblate",
+        ):
+            with self.subTest(repository=repository):
+                self.assertEqual(
+                    get_github_repository_identity(repository),
+                    ("github.com", "WeblateOrg/weblate"),
+                )
+
+        for repository in (
+            "file://github.com/WeblateOrg/weblate.git",
+            "ftp://github.com/WeblateOrg/weblate.git",
+            "local:",
+            "weblate://project/component",
+            "https://github.com/owner/repository/extra.git",
+            "https://github.com/../repository.git",
+        ):
+            with self.subTest(repository=repository):
+                self.assertIsNone(get_github_repository_identity(repository))
+
+    def test_clone_url_validation(self) -> None:
+        repository = {
+            "full_name": "WeblateOrg/weblate",
+            "clone_url": "https://github.com/WeblateOrg/weblate.git",
+        }
+        self.assertEqual(
+            get_github_repository_clone_url("github.com", repository),
+            repository["clone_url"],
+        )
+
+        repository["clone_url"] = "https://attacker.example/WeblateOrg/weblate.git"
+        self.assertEqual(
+            get_github_repository_clone_url("github.com", repository),
+            "https://github.com/WeblateOrg/weblate.git",
+        )
 
 
 def _make_credentials(